requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycparser==2.20
pylint==2.4.4
pyparsing==2.4.2
pytest==5.2.2
six==1.12.0
SQLAlchemy==1.3.19
typed-ast==1.4.1
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.11.2
zipp==0.6.0
py3-validate-email==0.1.4

flask-pytest
seleniumbase
python-dateutil